commit5:float input

+++ project1.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    long long int a = 0, b = 0;
-    scanf("%lld %lld", &a, &b);
+    float a = 0, b = 0;
+    scanf("%f %f", &a, &b);
     printf("%lld", a * b);
 }
